This is the new standard.
Patch Notes:
Implements a new log system to keep track of data.
Can now save the character to a new PDF file.
various other bug fixes and features

The install dependencies file is a beta. still a bit unstable. Ensure you are running the latest version of python and that it is installed in your PATH